Refactor table components to improve structure and readability; add multi-filters for products and enhance date filtering functionality. Update API endpoints to remove hardcoded URLs and streamline data handling in hooks. Introduce new styled components for better UI consistency and responsiveness.

+++ src/app/utils/functions/parseNumbers.jsx
@@ -5,7 +5,7 @@ export const parseNumber = (number) => {
 
         const parsed = number.includes(",") ?
             parseFloat(number.replace(",", ".")) :
-            parseInt(number)
+            parseFloat(number)
 
         return parsed
     }
